Separate all the classes used by Interpreter into their own TypeScript files
Separate classes used by Interpreter into their own TypeScript files.

* Add `ArrayCommand` class implementation in `src/ts/ArrayCommand.ts`.
* Add `CommandStack` class implementation in `src/ts/CommandStack.ts`.
* Add `ObjectCommand` class implementation in `src/ts/ObjectCommand.ts`.
* Add `ResponseData` class implementation in `src/ts/ResponseData.ts`.
* Add `Stack` class implementation in `src/ts/Stack.ts`.
* Add utility functions `mergeIn`, `pathRecur`, `isObject`, `addMetaData`, and `construct` in `src/ts/util.ts`.
* Import `Interpreter` class from `src/ts/Interpreter.ts` in `src/js/salliedforth.ts`.
* Update script tag to include the new `salliedforth-interpreter-test.js` file in `tests/index.html`.
* Update import statements to import the `Interpreter` class from `src/ts/Interpreter.ts` in the following test files:
  - `tests/js/salliedforth-interpreter-test.ts`
  - `tests/spec/browser.spec.ts`
  - `tests/spec/defining-words.spec.ts`
  - `tests/spec/flow-control.spec.ts`
  - `tests/spec/javascript-interop.spec.ts`
  - `tests/spec/maths.spec.ts`
  - `tests/spec/return-stack.spec.ts`
  - `tests/spec/stack-manipulation.spec.ts`

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/jococo/sallied-forth?shareId=XXXX-XXXX-XXXX-XXXX).
