+++ src/ts/CommandStack.ts
@@ -1,6 +1,5 @@
 import { Stack } from "./Stack";
 
-
 export class CommandStack extends Stack {
   constructor() {
     super("Command Stack");
