+++ tests/index.html
@@ -57,7 +57,7 @@
         }
       }
     </script>
-    <script src="js/salliedforth-interpreter-test.ts"></script>
+    <script src="js/salliedforth-interpreter-test.js"></script>
     <script src="spec/stack-manipulation.spec.ts"></script>
     <script src="spec/maths.spec.ts"></script>
     <script src="spec/flow-control.spec.ts"></script>
