+++ tests/spec/stack-manipulation.spec.ts
@@ -1,4 +1,4 @@
-import { Interpreter } from '../../src/js/salliedforth';
+import { Interpreter } from '../../src/ts/Interpreter';
 
 describe("Stack Manipulation", function() {
